requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
joblib==1.1.0
nltk==3.7
oauthlib==3.2.0
python-dotenv==0.20.0
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
textblob==0.17.1
tqdm==4.64.0
tweepy==4.10.0
urllib3==1.26.9
